+++ src/components/10-atoms/text/react/AXATextReactVersioned.jsx
@@ -1,7 +1,0 @@
-import { createElement } from 'react';
-import createAXAText from '../index.react';
-
-const customVersionSuffix = 'mypod';
-const AXAText = createAXAText(createElement, customVersionSuffix);
-
-export default AXAText;